Enh #3 - organize components, update auth routes in rails
