+++ client/src/features/Authenticate/AuthForm/Step3.jsx
@@ -6,7 +6,7 @@ import * as cn from 'classnames';
 import { createUseStyles, useTheme } from 'react-jss';
 import FormButton from './FormButton';
 import FormError from './FormError';
-import StyledInput from './StyledInput';
+import StyledInput from '../../../components/General/StyledInput';
 import useAuth from '../../../hooks/useAuth';
 import { styles } from '../../../utils/styles';
