+++ server/config/routes.rb
@@ -1,9 +1,8 @@
 Rails.application.routes.draw do
-  get 'authenticates/create'
   namespace :api do
     namespace :v1, defaults: { format: 'json' } do
       post 'login', to: 'authenticates#create'
-      post 'authenticates/handle', to: 'authenticates#handle'
+      post 'verify_handle', to: 'authenticates#verify_handle'
       resources :users
     end
   end
